refactor(retype): extract transformation rules and row converter
Split the retype command into rules.go and converter.go with matching
test files, keeping retype.go focused on command execution.
